requirements.txt
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
async-timeout==5.0.1
black==24.10.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jiter==0.7.1
mypy-extensions==1.0.0
openai==1.55.0
packaging==24.2
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
ptyprocess==0.7.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dotenv==1.0.1
redis==4.6.0
rq==2.1.0
sniffio==1.3.1
starlette==0.41.3
tomli==2.1.0
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.32.1
requests==2.28.1
psutil==5.8.0
google-genai==0.7.0
